+++ src/compose.ts
@@ -5,8 +5,7 @@ import { TypeBoxError, type TAnySchema, type TSchema } from '@sinclair/typebox'
 
 import { parseQuery, parseQueryFromURL } from './fast-querystring'
 
-// @ts-ignore
-import decodeURIComponent from 'fast-decode-uri-component'
+import { decode as decodeURIComponent } from 'deuri'
 
 import {
 	ELYSIA_REQUEST_ID,
@@ -777,8 +776,8 @@ export const composeHandler = ({
 							`else\n` +
 							`a${index}+=','\n` +
 							`let temp\n` +
-							`if(memory===-1)temp=decodeURIComponent(url.slice(start).replace(/\\+|%20/g,' '))\n` +
-							`else temp=decodeURIComponent(url.slice(start, memory).replace(/\\+|%20/g,' '))\n` +
+							`if(memory===-1)temp=decodeURIComponent(url.slice(start).replace(/\\+/g,' '))\n` +
+							`else temp=decodeURIComponent(url.slice(start, memory).replace(/\\+/g,' '))\n` +
 							`const charCode = temp.charCodeAt(0)\n` +
 							`if(charCode !== 91 && charCode !== 123)\n` +
 							`temp='"'+temp+'"'\n` +
@@ -801,10 +800,10 @@ export const composeHandler = ({
 							`if(a${index}===undefined)` +
 							`a${index}=[]\n` +
 							`if(memory===-1){` +
-							`a${index}.push(decodeURIComponent(url.slice(start)).replace(/\\+|%20/g,' '))\n` +
+							`a${index}.push(decodeURIComponent(url.slice(start)).replace(/\\+/g,' '))\n` +
 							`break` +
 							`}` +
-							`else a${index}.push(decodeURIComponent(url.slice(start, memory)).replace(/\\+|%20/g,' '))\n` +
+							`else a${index}.push(decodeURIComponent(url.slice(start, memory)).replace(/\\+/g,' '))\n` +
 							`memory=url.indexOf('&${key}=',memory)\n` +
 							`if(memory===-1) break\n` +
 							`}`
@@ -814,8 +813,8 @@ export const composeHandler = ({
 						`if(memory!==-1){` +
 						`const start=memory+${key.length + 2}\n` +
 						`memory=url.indexOf('&',start)\n` +
-						`if(memory===-1)a${index}=decodeURIComponent(url.slice(start).replace(/\\+|%20/g,' '))` +
-						`else a${index}=decodeURIComponent(url.slice(start,memory).replace(/\\+|%20/g,' '))` +
+						`if(memory===-1)a${index}=decodeURIComponent(url.slice(start).replace(/\\+/g,' '))` +
+						`else a${index}=decodeURIComponent(url.slice(start,memory).replace(/\\+/g,' '))` +
 						`if(a${index}!==undefined)` +
 						`try{` +
 						`a${index}=JSON.parse(a${index})` +
@@ -828,9 +827,9 @@ export const composeHandler = ({
 						`if(memory!==-1){` +
 						`const start=memory+${key.length + 2}\n` +
 						`memory=url.indexOf('&',start)\n` +
-						`if(memory===-1)a${index}=decodeURIComponent(url.slice(start).replace(/\\+|%20/g,' '))\n` +
+						`if(memory===-1)a${index}=decodeURIComponent(url.slice(start).replace(/\\+/g,' '))\n` +
 						`else{` +
-						`a${index}=decodeURIComponent(url.slice(start,memory).replace(/\\+|%20/g,' '))`
+						`a${index}=decodeURIComponent(url.slice(start,memory).replace(/\\+/g,' '))`
 
 					if (anyOf)
 						fnLiteral +=
@@ -843,8 +842,8 @@ export const composeHandler = ({
 							`if(first)first=false\n` +
 							`else deepMemory = url.indexOf('&', start)\n` +
 							`let value\n` +
-							`if(deepMemory===-1)value=decodeURIComponent(url.slice(start).replace(/\\+|%20/g,' '))\n` +
-							`else value=decodeURIComponent(url.slice(start, deepMemory).replace(/\\+|%20/g,' '))\n` +
+							`if(deepMemory===-1)value=decodeURIComponent(url.slice(start).replace(/\\+/g,' '))\n` +
+							`else value=decodeURIComponent(url.slice(start, deepMemory).replace(/\\+/g,' '))\n` +
 							`const vStart=value.charCodeAt(0)\n` +
 							`const vEnd=value.charCodeAt(value.length - 1)\n` +
 							`if((vStart===91&&vEnd===93)||(vStart===123&&vEnd===125))\n` +