+++ src/utils.ts
@@ -1242,9 +1242,9 @@ export const traceBackMacro = (
 					})
 				}
 			}
-
-			delete property[key as unknown as keyof typeof extension]
 		}
+
+		delete property[key as unknown as keyof typeof extension]
 	}
 }
 
@@ -1272,7 +1272,7 @@ export const createMacroManager =
 			}
 
 		// @ts-expect-error this is available in macro v2
-		if(stackName === "resolve") {
+		if (stackName === 'resolve') {
 			type = {
 				...type,
 				subType: 'resolve'